Moved CodecUtil into util
