+++ src/main/java/fr/estecka/variantscit/assetgen/GeneratorPresets.java
@@ -6,7 +6,7 @@ import java.util.regex.Pattern;
 import net.minecraft.resources.Identifier;
 import net.minecraft.server.packs.resources.ResourceManager;
 import com.mojang.serialization.Codec;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.VariantsCitMod;
 import fr.estecka.variantscit.reload.ModuleDefinition;
