+++ src/main/java/fr/estecka/variantscit/assetgen/IAssetGenerator.java
@@ -6,7 +6,7 @@ import java.util.List;
 import net.minecraft.resources.Identifier;
 import net.minecraft.server.packs.resources.IoSupplier;
 import com.mojang.serialization.Codec;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 
 public interface IAssetGenerator
 {
