+++ src/main/java/fr/estecka/variantscit/assetgen/TemplateRepository.java
@@ -10,7 +10,7 @@ import net.minecraft.server.packs.resources.Resource;
 import net.minecraft.server.packs.resources.ResourceManager;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.DataResult;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.VariantsCitMod;
 import fr.estecka.variantscit.format.Substitution;
