+++ src/main/java/fr/estecka/variantscit/assetgen/TemplatedAssetGenerator.java
@@ -9,7 +9,7 @@ import org.apache.commons.lang3.exception.ExceptionUtils;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.VariantsCitMod;
 
 public record TemplatedAssetGenerator(
