+++ src/main/java/fr/estecka/variantscit/itemdata/containers/ComponentContainer.java
@@ -2,7 +2,7 @@ package fr.estecka.variantscit.itemdata.containers;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.nbt.Tag;
