+++ src/main/java/fr/estecka/variantscit/itemdata/transforms/SuccessiveTransform.java
@@ -3,7 +3,7 @@ package fr.estecka.variantscit.itemdata.transforms;
 import java.util.List;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.itemdata.containers.IDataContainer;
 
 public record SuccessiveTransform(IDataTransform[] subTransforms)
