+++ src/main/java/fr/estecka/variantscit/itemdata/transforms/impl/RegexTransform.java
@@ -7,7 +7,7 @@ import org.apache.commons.lang3.exception.ExceptionUtils;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.VariantsCitMod;
 import fr.estecka.variantscit.itemdata.transforms.IStringTransform;
