+++ src/main/java/fr/estecka/variantscit/itemdata/transforms/impl/TestTransform.java
@@ -1,7 +1,7 @@
 package fr.estecka.variantscit.itemdata.transforms.impl;
 
 import com.mojang.serialization.MapCodec;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.itemdata.containers.IDataContainer;
 import fr.estecka.variantscit.itemdata.transforms.IDataTransform;
 import fr.estecka.variantscit.itemdata.transforms.SuccessiveTransform;
