+++ src/main/java/fr/estecka/variantscit/modules/impl/AxolotlBucketModule.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import net.minecraft.resources.Identifier;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.modules.libraries.VariantLibrary;
 import fr.estecka.variantscit.reload.IUnbakedModule;
 import fr.estecka.variantscit.modules.libraries.GenericBakedModule;
