+++ src/main/java/fr/estecka/variantscit/modules/impl/ComponentThresholdModule.java
@@ -11,7 +11,7 @@ import org.jetbrains.annotations.Nullable;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.commands.CommandLogger;
 import fr.estecka.variantscit.itemdata.transforms.impl.NbtPath;
 import fr.estecka.variantscit.modules.cache.CacheKeySet;
