+++ src/main/java/fr/estecka/variantscit/modules/libraries/LinearModuleBaker.java
@@ -2,7 +2,7 @@ package fr.estecka.variantscit.modules.libraries;
 
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.util.CodecUtil;
 import fr.estecka.variantscit.reload.IUnbakedModule;
 import net.minecraft.resources.Identifier;
